Split service configuration for desktop and browser
Fixes #128222

We only support using configured TS versions on desktop. In browsers we always use the bundled version
